refactor : modal 컴포넌트 폴더구조 변경 - 컴포넌트별 폴더 생성
