+++ components/src/lib/ModalContent/ModalContent.tsx
@@ -1,0 +1,7 @@
+interface ModalContentProps {
+  children: React.ReactNode;
+}
+
+export default function ModalContent({ children }: ModalContentProps) {
+  return <div>{children}</div>;
+}
