+++ components/src/lib/ModalFooter/ModalFooter.tsx
@@ -1,0 +1,7 @@
+interface ModalFooterProps {
+  children: React.ReactNode;
+}
+
+export default function ModalFooter({ children }: ModalFooterProps) {
+  return <footer>{children}</footer>;
+}
